requirements.txt
# -------------------------------
# Core dependencies
# -------------------------------
torch==2.7.1+cu126
torchvision==0.22.1
torchaudio==2.7.1
--index-url https://download.pytorch.org/whl/cu126

# -------------------------------
# Hugging Face ecosystem
# -------------------------------
transformers==4.41.0
huggingface-hub==1.0.0rc6
datasets==4.4.0
accelerate>=0.29.0
evaluate==0.4.1

# -------------------------------
# ML & Scientific stack
# -------------------------------
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
scikit-learn==1.3.2
imbalanced-learn==0.11.0
tqdm==4.67.1

# -------------------------------
# Vision / Image
# -------------------------------
opencv-python-headless==4.10.0.82
Pillow>=10.3.0
dlib>=19.24.2

# -------------------------------
# Utilities & Misc
# -------------------------------
matplotlib>=3.8.4
aifactory>=0.1.0
cmake>=3.22.0
build-essential>=12.9